Тип суммы,  обработка возврата, проверка темы

+++ app/src/main/java/com/example/sandboxbank/App/ui/applycredit/entity/ApplyCreditEntity.kt
@@ -1,6 +1,8 @@
 package com.example.sandboxbank.App.ui.applycredit.entity
 
+import java.math.BigDecimal
+
 data class ApplyCreditEntity(
-    val sum: Long,
+    val sum: BigDecimal,
     val time: Int,
 )
